Split tool_actions into service-specific modules
Reorganize tracer_actions.py into:
- actions_aws_batch.py: AWS Batch job actions
- actions_tracer_api.py: Tracer API client actions
- actions_tracer_web.py: Tracer Web App actions

Update imports and maintain backward compatibility via __init__.py.
